requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
beautifultable==0.8.0
boto==2.49.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-appconf==1.0.3
django-autoslug==1.9.4
django-bootstrap-form==3.4
django-braces==1.13.0
django-comments-xtd==2.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-filter==2.1.0
django-froala-editor==3.0.5
django-google-maps==0.11.0
django-hvad==1.8.0
django-jalali==3.1.0
django-libs==2.0.2
django-map-widgets==0.2.2
django-markdown-deux==1.0.5
django-materialize-css==0.0.1
django-model-utils==3.1.2
django-modelcluster==4.4
django-multiselectfield==0.1.8
django-phonenumber-field==2.4.0
django-social-share==1.3.2
django-storages==1.7.1
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
docutils==0.14
draftjs-exporter==2.1.5
email-reply-parser==0.5.9
fandogh-cli==1.15.1
filelock==3.0.12
html5lib==1.0.1
idna==2.8
importlib-metadata==0.18
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jdatetime==3.2.0
jedi==0.13.3
lxml==4.4.1
Markdown==3.1.1
markdown2==2.3.8
ndg-httpsclient==0.5.1
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
phonenumbers==8.10.11
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.12.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
tzlocal==2.0.0
Unidecode==1.0.23
urllib3==1.24.3
virtualenv==16.6.1
wagtail==2.5.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
websockets==3.2
Willow==1.1
zipp==0.5.1
